+++ lib/dotup_device_simulator.dart
@@ -1,3 +1,5 @@
 library devicesimulator;
 
 export 'src/device_simulator.dart';
+export 'src/device_spec_list.dart';
+export 'src/device_specification.dart';